Initial inclusion of ffmpeg-chromium
Summary:
Chromium's ffmpeg for opera and vivaldi

NOTE: This is a temporary package until such a time when chromium enters the repo. When it does, this
package will be deprecated and replaced by the `chromium-ffmpeg` subpackage.

Signed-off-by: Joey Riches <josephriches@gmail.com>

Test Plan: Remove existing vendored `libffmpeg.so` in opera/vivaldi packages and add `ffmpeg-chromium` to builddeps, test with a H.264 video.

Reviewers: #triage_team

Subscribers: JoshStrobl

Differential Revision: https://dev.solus-project.com/D1215